Add -D/--dpms option to power off displays when idle
This option makes use of the new wlr-output-power-management protocol to
power off displays when swaylock is idle (i.e. when the unlock
indicator, if enabled, is not shown).

This approach provides far more reliable DPMS control than the currently
recommended solution, which is to use something above swaylock (e.g.
swayidle) to coordinate swaylock invocation with a separate swaymsg
invocation to control DPMS. Since swayidle is purely input-based and
cannot inspect swaylock's state, it will inevitably either power up the
screens when it shouldn't or fail to when it should. For example, if
swayidle is set to turn on the screens on "resume," it will do so even
if the only input was a mouse bump, modifier key press, or something
else that swaylock ignores completely.

Signed-off-by: Thomas Hebb <tommyhebb@gmail.com>